Fix for new intellij update
An intellij update broke support for a maven feature, so that when opening this project with Intellij caused the bytecode version not to be set to 11, but to 1.5. This change to the old way of setting the java version should fix this so that the bytecode version is correctly set.